[refactor/#353] unit 경로로 이동
